requirements.txt
Django==4.2.9
image==1.5.33
sqlparse==0.4.4
psycopg2-binary==2.9.9
gunicorn==19.7.1
premailer==3.10.0
XlsxWriter==3.2.0
django-simple-captcha==0.6.0